Avoid unnecessary use of eval (insecure)
The appropriate way to convert a hexadecimal string to integer is by using the built-in functionality of `int`. The `eval` function is insecure; if the input could possibly come from outside the program, no matter how indirectly, this creates a risk of an arbitrary code execution exploit. It's also much less efficient.

Also renamed the function parameter so as to avoid shadowing the built-in name `str`.